Moving scan retry logic into a separate class. This will allow us to use the retry functionality in other scenarios, specifically observer based responses as described in #703.
